bash rc's all cleaned up

+++ .bash_aliases
@@ -1,17 +1,9 @@
-#!/bin/sh
+echo " BASH ALIASES"
 
-#check for mac
-if [ $(uname -s) == "Darwin" ]
-then
-    #echo "Using mac!"
-    alias ls='ls -G'
-    alias ll='ls -al'
-
-    if [ -e /usr/local/bin/virtualenvwrapper.sh ]
-    then
-        source /usr/local/bin/virtualenvwrapper.sh
-    fi
-fi
+alias ls='ls -G'
+alias la='ls -A'
+alias l='ls -CF'
+alias ll='ls -al'
 
 alias grep='grep --color=auto'
 alias fgrep='fgrep --color=auto'
@@ -23,10 +15,19 @@ alias ....='cd ../../..'
 
 alias weather="curl -4 http://wttr.in/sanfrancisco"
 
-if [ -f /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl ]; then
-  ln -s /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl /usr/local/bin/sublime
-fi
+# Check for OSX/macOS
+if [ $(uname -s) == "Darwin" ]
+then
+    echo 'using mac'
+    if [ ! -f /usr/local/bin/sublime -a -f /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl ]; then
+      ln -s /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl /usr/local/bin/sublime
+    fi
 
+    if [ -e /usr/local/bin/virtualenvwrapper.sh ]
+    then
+        source /usr/local/bin/virtualenvwrapper.sh
+    fi
+fi
 
 mkdircdfunc() {
     #do things with parameters like $1 such as
